refactor: 💡 renamed component to columnsSelector
